Large refactor (#27)
* AI module

* ARC module

* BASE module

* CNT module

* DB module

* DSP module

* DVD module

* EXI module

* FS module

* GX module

* IPC module

* MEM module

* NAND module

* NdevExi2AD module

* OS module

* RFL work

* Refactor RFL_System

* Refactor RFL_NANDLoader

* WIP RFL_NANDAccess

* Refactor RFL_NANDAccess

* Change `scTmpSize` to a macro

* RFL first pass

* RFL second pass: Split public/internal API

* Refactor RFL_Model, attempt all functions

* SC module

* TPL module

* TPL: Fix decl argument names

* USB module

* VF module

* Update vi.h

+++ src/revolution/OS/OSAlloc.c
@@ -1,4 +1,4 @@
-#include <OS.h>
+#include <revolution/OS.h>
 
 typedef struct OSHeapDescriptor {
     s32 size;                    // at 0x0
@@ -179,7 +179,7 @@ s32 OSCreateHeap(void* start, void* end) {
         OSHeapCell* cell = (OSHeapCell*)start;
         OSHeapDescriptor* hd = &HeapArray[handle];
         if (hd->size < 0) {
-            hd->size = (u32)end - (u32)start;
+            hd->size = (uintptr_t)end - (uintptr_t)start;
 
             cell->prev = NULL;
             cell->next = NULL;
